:fire: move usecase/production to infra/repository
