+++ infra/repository/group.go
@@ -1,4 +1,4 @@
-package production
+package repository
 
 import (
 	"errors"
