+++ infra/repository/room.go
@@ -1,4 +1,4 @@
-package production
+package repository
 
 import (
 	"time"
