+++ main.go
@@ -10,8 +10,8 @@ import (
 
 	"github.com/traPtitech/knoQ/domain"
 	"github.com/traPtitech/knoQ/infra/db"
+	"github.com/traPtitech/knoQ/infra/repository"
 	"github.com/traPtitech/knoQ/infra/traq"
-	"github.com/traPtitech/knoQ/usecase/production"
 	"github.com/traPtitech/knoQ/utils"
 	"golang.org/x/oauth2"
 
@@ -69,7 +69,7 @@ func main() {
 		},
 		URL: "https://q.trap.jp/api/v3",
 	}
-	repo := &production.Repository{
+	repo := &repository.Repository{
 		GormRepo: gormRepo,
 		TraQRepo: traqRepo,
 	}
